add OzTiVo style datalist generator to german site
This adds a new file datalist.xml.gz to the xmltv export which contains
the last modified timestamps of the individual guide data files.
Once this file is rolled out the grabber only has to request changed
files without asking the server for each file on speculation.
This makes it relatively cheap to grab the whole guide time span because
the grabber caches more aggressive.

The next step is to extend the xmltv grabber to work with both the old
and the new file. It should conditionally report "preferred mode == allatonce"
to the consuming application if it is using the new format.
With the new grabber code rolled out every instace of the grabber can be
switched to/from the new format by simply setting the sites default url
to the old or new file.