Add DVB service pointer to map DVB services to guide channels
File format is a proposed at http://wiki.xmltv.org/index.php/NoLineupProposal
Both ETSI TS 102 323 and the Eutelsat implemtation guide handle the transport_id
as optional for the differentiation of DVB services, lets see how it works out
in practice.
The format allows to specify a transport_id for cases where two services on
neighbouring transports in the same original_network share the same service_id
even though that is a bug in the implementation of the DVB network and has been
fixed in the past.
Notice that it is expected to see collisions in the service_id domain when not
taking the original_network_id into account!